lambda-function/requirements.txt
boto3==1.34.34
Pillow==10.1.0